agent-camel-v2/requirements.txt
# Core dependencies
camel-ai>=0.1.0
camel-ai[web_tools]>=0.1.0
langchain>=0.0.300
llama-index>=0.8.0

# LLM providers
openai>=1.0.0
anthropic>=0.5.0

# Data storage
chromadb>=0.4.0
pinecone-client>=2.2.0
redis>=4.0.0
psycopg2-binary>=2.9.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
uuid>=1.30
requests>=2.30.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0